refactor: extract commitment utils
